+++ tests/Unit/Api/Role/ListTest.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Redmine\Tests\Unit\Api\Role;
+
+use PHPUnit\Framework\TestCase;
+use Redmine\Api\Role;
+use Redmine\Client\Client;
+
+/**
+ * @covers \Redmine\Api\Role::list
+ */
+class ListTest extends TestCase
+{
+    public function testListWithoutParametersReturnsResponse()
+    {
+        // Test values
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->stringStartsWith('/roles.json')
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new Role($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->list());
+    }
+
+    public function testListWithParametersReturnsResponse()
+    {
+        // Test values
+        $parameters = ['not-used'];
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->logicalAnd(
+                    $this->stringStartsWith('/roles.json'),
+                    $this->stringContains('not-used')
+                )
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new Role($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->list($parameters));
+    }
+}